Prelim thermistor reading.

+++ Test_Fixture_3/src/main.h
@@ -20,7 +20,7 @@ void SetChipConfigurations(uint8_t localConfig[][6]);
 void ConfigureDischarge(uint8_t chip, uint16_t cells);
 void ConfigureCOMMRegisters(uint8_t numChips, uint8_t dataToWrite[][3], uint8_t commOutput [][6]);
 void ConfigureCOMMRegisters2(uint8_t numChips, uint8_t datatToWrite[][3], uint8_t commOutput[][6]);
-void GetThermistor(uint8_t thermisters[]);
+void GetThermistor(float thermisters[]);
 
 
 #endif
